requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
python-3.9
apache-airflow==2.5.0
apache-airflow-providers-http==3.2.0
apache-airflow-providers-postgres==3.2.0